+++ src/main/java/Application.java
@@ -1,9 +1,11 @@
 import manager.GameManager;
+import view.InputView;
 
 public class Application {
 
     public static void main(String[] args) {
-        GameManager gameManager = new GameManager("1");
-        gameManager.startGame();
+        GameManager gameManager = new GameManager();
+        String gameRoomName = InputView.inputGameRoomName();
+        gameManager.startGame(gameRoomName);
     }
 }
