+++ src/main/java/domain/PieceType.java
@@ -1,4 +1,4 @@
-package domain.piece;
+package domain;
 
 public enum PieceType {
