+++ src/main/java/domain/board/Board.java
@@ -1,8 +1,12 @@
 package domain.board;
 
 import domain.piece.Piece;
-import domain.piece.PieceType;
-import domain.piece.Team;
+import domain.PieceType;
+import domain.Team;
+import domain.point.Direction;
+import domain.point.Path;
+import domain.point.PathFinder;
+import domain.point.Point;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
