+++ src/main/java/domain/piece/Piece.java
@@ -1,7 +1,9 @@
 package domain.piece;
 
+import domain.PieceType;
+import domain.Team;
 import domain.board.PieceVisibleBoard;
-import domain.board.Point;
+import domain.point.Point;
 import java.util.List;
 
 public abstract class Piece {
