+++ src/main/java/domain/piece/Po.java
@@ -1,8 +1,10 @@
 package domain.piece;
 
-import domain.board.Direction;
+import domain.PieceType;
+import domain.Team;
 import domain.board.PieceVisibleBoard;
-import domain.board.Point;
+import domain.point.Direction;
+import domain.point.Point;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
