+++ src/main/java/domain/piece/Wang.java
@@ -1,6 +1,8 @@
 package domain.piece;
 
-import domain.board.Direction;
+import domain.PieceType;
+import domain.Team;
+import domain.point.Direction;
 import java.util.Arrays;
 import java.util.List;
