+++ src/main/java/view/InputView.java
@@ -1,7 +1,7 @@
 package view;
 
-import domain.board.Point;
-import domain.piece.Team;
+import domain.Team;
+import domain.point.Point;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
@@ -10,7 +10,11 @@ import util.ErrorHandler;
 public class InputView {
 
     private static final Scanner scanner = new Scanner(System.in);
-    private static final String MOVE_COMMAND_INPUT = "move";
+
+    public static String inputGameRoomName() {
+        System.out.printf("%n입장할 게임방 이름을 입력해주세요.%n> ");
+        return scanner.nextLine();
+    }
 
     public static SangMaOrderCommand inputSangMaOrder(Team team) {
         return ErrorHandler.retryUntilSuccessWithReturn(() -> {
@@ -20,7 +24,7 @@ public class InputView {
                     2. 상마마상
                     3. 마상상마
                     4. 마상마상
-                    """, team.title());
+                    >\s""", team.title());
             String input = scanner.nextLine();
             return SangMaOrderCommand.from(input);
         });
@@ -33,7 +37,7 @@ public class InputView {
                     > move: 기물 이동
                     > status: 현재 점수 출력
                     > end: 게임 종료
-                    """);
+                    >\s""");
             String input = scanner.nextLine();
             return Command.from(input.strip());
         });
@@ -44,7 +48,7 @@ public class InputView {
             System.out.printf("""
                     %n현재 턴: %s나라
                     이동할 기물의 현재 위치와 이동할 위치를 입력해주세요. (예: 7,1 7,2)
-                    """, team.title());
+                    >\s""", team.title());
             String input = scanner.nextLine();
 
             List<String> parsed = Arrays.stream(input.split(" ", -1)).toList();
